Refactor to use common chart component
